Move to `ExtensionTab`
