+++ app/src/main/java/eu/kanade/presentation/browse/ExtensionsScreen.kt
@@ -1,6 +1,5 @@
 package eu.kanade.presentation.browse
 
-import android.content.Context
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.combinedClickable
@@ -56,7 +55,6 @@ import eu.kanade.tachiyomi.extension.model.InstallStep
 import eu.kanade.tachiyomi.ui.browse.extension.ExtensionUiModel
 import eu.kanade.tachiyomi.ui.browse.extension.ExtensionsScreenModel
 import eu.kanade.tachiyomi.util.system.LocaleHelper
-import eu.kanade.tachiyomi.util.system.isPackageInstalled
 import eu.kanade.tachiyomi.util.system.launchRequestPackageInstallsPermission
 import kotlinx.collections.immutable.persistentListOf
 import tachiyomi.i18n.MR
@@ -250,7 +248,6 @@ private fun ExtensionContent(
                             }
                         }
                     },
-                    context = context,
                 )
             }
         }
@@ -281,43 +278,16 @@ private fun ExtensionItem(
     onClickItemAction: (Extension) -> Unit,
     onClickItemSecondaryAction: (Extension) -> Unit,
     modifier: Modifier = Modifier,
-    context: Context,
 ) {
     val (extension, installStep) = item
-    var showUninstallConfirmation by remember { mutableStateOf(false) }
-
-    val internalOnLongClickItem: () -> Unit = {
-        if (context.isPackageInstalled(extension.pkgName)) {
-            onLongClickItem(extension)
-        } else {
-            showUninstallConfirmation = true
-        }
-    }
-
-    if (showUninstallConfirmation) {
-        ExtensionUninstallConfirmation(
-            extensionName = extension.name,
-            onClickConfirm = {
-                onLongClickItem(extension)
-                showUninstallConfirmation = false
-            },
-            onClickDismiss = {
-                showUninstallConfirmation = false
-            },
-            onDismissRequest = {
-                showUninstallConfirmation = false
-            },
-        )
-    }
-
     BaseBrowseItem(
         modifier = modifier
             .combinedClickable(
                 onClick = { onClickItem(extension) },
-                onLongClick = internalOnLongClickItem,
+                onLongClick = { onLongClickItem(extension) },
             ),
         onClickItem = { onClickItem(extension) },
-        onLongClickItem = internalOnLongClickItem,
+        onLongClickItem = { onLongClickItem(extension) },
         icon = {
             Box(
                 modifier = Modifier
@@ -565,34 +535,6 @@ private fun ExtensionTrustDialog(
         dismissButton = {
             TextButton(onClick = onClickDismiss) {
                 Text(text = stringResource(MR.strings.ext_uninstall))
-            }
-        },
-        onDismissRequest = onDismissRequest,
-    )
-}
-
-@Composable
-private fun ExtensionUninstallConfirmation(
-    extensionName: String,
-    onClickConfirm: () -> Unit,
-    onClickDismiss: () -> Unit,
-    onDismissRequest: () -> Unit,
-) {
-    AlertDialog(
-        title = {
-            Text(text = stringResource(MR.strings.ext_confirm_remove))
-        },
-        text = {
-            Text(text = stringResource(MR.strings.remove_private_extension_message, extensionName))
-        },
-        confirmButton = {
-            TextButton(onClick = onClickConfirm) {
-                Text(text = stringResource(MR.strings.ext_remove))
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onClickDismiss) {
-                Text(text = stringResource(MR.strings.action_cancel))
             }
         },
         onDismissRequest = onDismissRequest,
