+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/extension/ExtensionsTab.kt
@@ -1,8 +1,14 @@
 package eu.kanade.tachiyomi.ui.browse.extension
 
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.platform.LocalContext
 import cafe.adriel.voyager.navigator.LocalNavigator
 import cafe.adriel.voyager.navigator.currentOrThrow
 import eu.kanade.presentation.browse.ExtensionScreen
@@ -12,6 +18,7 @@ import eu.kanade.presentation.more.settings.screen.browse.ExtensionReposScreen
 import eu.kanade.tachiyomi.extension.model.Extension
 import eu.kanade.tachiyomi.ui.browse.extension.details.ExtensionDetailsScreen
 import eu.kanade.tachiyomi.ui.webview.WebViewScreen
+import eu.kanade.tachiyomi.util.system.isPackageInstalled
 import kotlinx.collections.immutable.persistentListOf
 import tachiyomi.i18n.MR
 import tachiyomi.presentation.core.i18n.stringResource
@@ -22,6 +29,8 @@ fun extensionsTab(
 ): TabContent {
     val navigator = LocalNavigator.currentOrThrow
     val state by extensionsScreenModel.state.collectAsState()
+    val context = LocalContext.current
+    val showUninstallConfirmation = remember { mutableStateOf<Extension?>(null) }
 
     return TabContent(
         titleRes = MR.strings.label_extensions,
@@ -45,7 +54,13 @@ fun extensionsTab(
                 onLongClickItem = { extension ->
                     when (extension) {
                         is Extension.Available -> extensionsScreenModel.installExtension(extension)
-                        else -> extensionsScreenModel.uninstallExtension(extension)
+                        else -> {
+                            if (context.isPackageInstalled(extension.pkgName)) {
+                                extensionsScreenModel.uninstallExtension(extension)
+                            } else {
+                                showUninstallConfirmation.value = extension
+                            }
+                        }
                     }
                 },
                 onClickItemCancel = extensionsScreenModel::cancelInstallUpdateExtension,
@@ -68,6 +83,50 @@ fun extensionsTab(
                 onUpdateExtension = extensionsScreenModel::updateExtension,
                 onRefresh = extensionsScreenModel::findAvailableExtensions,
             )
+
+            if (showUninstallConfirmation.value != null) {
+                ExtensionUninstallConfirmation(
+                    extensionName = showUninstallConfirmation.value!!.name,
+                    onClickConfirm = {
+                        extensionsScreenModel.uninstallExtension(showUninstallConfirmation.value!!)
+                        showUninstallConfirmation.value = null
+                    },
+                    onClickDismiss = {
+                        showUninstallConfirmation.value = null
+                    },
+                    onDismissRequest = {
+                        showUninstallConfirmation.value = null
+                    },
+                )
+            }
         },
+    )
+}
+
+@Composable
+private fun ExtensionUninstallConfirmation(
+    extensionName: String,
+    onClickConfirm: () -> Unit,
+    onClickDismiss: () -> Unit,
+    onDismissRequest: () -> Unit,
+) {
+    AlertDialog(
+        title = {
+            Text(text = stringResource(MR.strings.ext_confirm_remove))
+        },
+        text = {
+            Text(text = stringResource(MR.strings.remove_private_extension_message, extensionName))
+        },
+        confirmButton = {
+            TextButton(onClick = onClickConfirm) {
+                Text(text = stringResource(MR.strings.ext_remove))
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onClickDismiss) {
+                Text(text = stringResource(MR.strings.action_cancel))
+            }
+        },
+        onDismissRequest = onDismissRequest,
     )
 }
